// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/reportes/reporteInformacionOficialBean.ts
export class ReporteInformacionOficialBean {
    ubigeo: string;
    codigoEleccion: string;
    codigoAmbitoElectoral: string;
    codigoCentroComputo: string;
    nivelUbigeoUno: string;
    nivelUbigeoDos: string;
    nivelUbigeoTres: string;
    electoresHabiles: number;
    ciudadanoVotaron: number;
    ciudadanoNoVotaron: number;
    ausentismoCifras: number;
    porcentajeParticipacion: number;
    porcentajeAusentismo: number;
    idCentroComputo: number;
    ambitoGeografico: string;
}